add rapids-cli (#55)
Adds https://github.com/rapidsai/rapids-cli, which we recently started
publishing packages for:

* https://github.com/rapidsai/rapids-cli/issues/89
* https://github.com/rapidsai/rapids-cli/issues/90

Similar to `pynvjitlink` (which is already included in this repo), this
project does ad hoc releases not aligned to the RAPIDS release schedule,
and uses a 3-part version number (e.g., the latest release was `0.1.2`).

Adding this to 25.06 because it was included in the 25.06
`rapidsai/docker` images: https://github.com/rapidsai/docker/pull/764

While I'm touching this repo, also updated all the `pre-commit` hooks
with `pre-commit autoupdate`.